Ensure that names in make_index_unique are unique
Fixes https://github.com/theislab/anndata/issues/344
